cgroup/thp:Add cgroup_thp_control feature
category: feature
bugzilla: NA
CVE : NA

---------------------------

Using THP may promote the performance of memory, but increase memory
footprint. Applications may use madvise to decrease footprint, but
not all applications support using madvise, and it takes much costs
to re-code all the applications. And we notice container becomes more
and more popular to manage a set of tasks.

So add support for cgroup to control THP behaviour will provide much
convenience, administrator may only enable THP for important containers,
and disable it for other containers. Then we can enjoy the high performance
of THP while minimize memory footprint without re-coding any application.
See /sys/fs/cgroup/memory.thp_enabled or
/sys/fs/cgroup/[cgroup ID]/memory.thp_enabled. It's semantic is just like
/sys/kernel/mm/transparent_hugepage/enabled.

Also add a knode to disable this feature in case of emergency. See
/sys/kernel/mm/transparent_hugepage/cgroup_enabled. It is disable
by default, and this feature is closed. It can be set to enable,
and then this feature is on.

Cgroupv1 is used for many distributions, so this feature is based
on it.

To enable the cgroup_thp_control feature on x86/arm64

Signed-off-by: YangYang <yang.yang29@zte.com.cn>
Signed-off-by: wenzhiwei <wenzhiwei@kylinos.cn>
Signed-off-by: lilong <lilong@kylinos.cn>